+++ instrumentation/pg/lib/opentelemetry/instrumentation/pg/instrumentation.rb
@@ -26,6 +26,7 @@ module OpenTelemetry
 
         option :peer_service, default: nil, validate: :string
         option :db_statement, default: :include, validate: %I[omit include obfuscate]
+        option :obfuscation_limit, default: 2000, validate: :integer
 
         private