+++ instrumentation/all/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Release History: opentelemetry-instrumentation-all
 
+### v0.39.1 / 2023-06-05
+
+* FIXED: Use latest bug fix version for all dependencies
+
+### v0.39.0 / 2023-06-02
+
+* BREAKING CHANGE: Separate logical MySQL host from connected host
+* ADDED: Separate logical MySQL host from connected host
+
+### v0.38.0 / 2023-05-31
+
+* BREAKING CHANGE: Add database name for trilogy traces
+
+* ADDED: Add database name for trilogy traces
+
+### v0.37.0 / 2023-05-25
+
+* ADDED: Add config[:obfuscation_limit] to pg and mysql2
+* ADDED: Add Obfuscation Limit Option to Trilogy
+
+
 ### v0.36.0 / 2023-05-18
 
 * ADDED: GraphQL instrumentation: support new tracing API (#453)
@@ -14,8 +35,8 @@
 
 * BREAKING CHANGE: Drop support for EoL Ruby 2.7
 
-* ADDED: Drop support for EoL Ruby 2.7 
-* ADDED: Add Grape instrumentation 
+* ADDED: Drop support for EoL Ruby 2.7
+* ADDED: Add Grape instrumentation
 
 ### v0.33.0 / 2023-03-15