+++ instrumentation/graphql/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release History: opentelemetry-instrumentation-graphql
 
+### v0.26.2 / 2023-06-05
+
+* FIXED: Base config options 
+* FIXED: GraphQL resolve_type_lazy 
+
+### v0.26.1 / 2023-05-30
+
+* FIXED: GraphQL tracing
+
 ### v0.26.0 / 2023-05-17
 
 * BREAKING CHANGE: GraphQL instrumentation: support new tracing API