+++ instrumentation/trilogy/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Release History: opentelemetry-instrumentation-trilogy
 
+### v0.56.1 / 2023-06-05
+
+* FIXED: Base config options 
+
+### v0.56.0 / 2023-06-02
+
+* BREAKING CHANGE: Separate logical MySQL host from connected host 
+
+* ADDED: Separate logical MySQL host from connected host 
+
+### v0.55.1 / 2023-06-01
+
+* FIXED: Regex non-match with obfuscation limit (issue #486) 
+
+### v0.55.0 / 2023-05-31
+
+* BREAKING CHANGE: Add database name for trilogy traces 
+
+* ADDED: Add database name for trilogy traces 
+
+### v0.54.0 / 2023-05-25
+
+* ADDED: Add Obfuscation Limit Option to Trilogy 
+
 ### v0.53.0 / 2023-04-17
 
 * BREAKING CHANGE: Drop support for EoL Ruby 2.7